requirements.txt
Django==2.1.2
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.8.2
djangorestframework-recursive==0.1.2
Pillow==5.3.0
psycopg2-binary==2.7.5
pytz==2018.5
